requirements.txt
streamlit
openai
faiss-cpu
langchain
pyMuPDF
